Split DuidTests class into multiple concern based class partials
