+++ src/NetDuid.Tests/Duid.IParsableTests.cs
@@ -1,0 +1,4 @@
+﻿namespace NetDuid.Tests
+{
+    public partial class DuidTests { }
+}
